+++ node/actors/bft/src/tests/mod.rs
@@ -1,0 +1,54 @@
+use crate::testonly::{Behavior, Network, Test};
+use zksync_concurrency::{ctx, time};
+use zksync_consensus_roles::validator;
+
+mod twins;
+
+async fn run_test(behavior: Behavior, network: Network) {
+    tokio::time::pause();
+    let _guard = zksync_concurrency::testonly::set_timeout(time::Duration::seconds(30));
+    zksync_concurrency::testonly::abort_on_panic();
+    let ctx = &ctx::test_root(&ctx::RealClock);
+
+    const NODES: usize = 11;
+    let mut nodes = vec![(behavior, 1u64); NODES];
+    // validator::threshold(NODES) will calculate required nodes to validate a message
+    // given each node weight is 1
+    let honest_nodes_amount = validator::quorum_threshold(NODES as u64) as usize;
+    for n in &mut nodes[0..honest_nodes_amount] {
+        n.0 = Behavior::Honest;
+    }
+    Test {
+        network,
+        nodes,
+        blocks_to_finalize: 10,
+    }
+    .run(ctx)
+    .await
+    .unwrap()
+}
+
+#[tokio::test]
+async fn honest_real_network() {
+    run_test(Behavior::Honest, Network::Real).await
+}
+
+#[tokio::test]
+async fn offline_real_network() {
+    run_test(Behavior::Offline, Network::Real).await
+}
+
+/// Not being able to propose a block shouldn't cause a deadlock.
+#[tokio::test]
+async fn non_proposing_leader() {
+    zksync_concurrency::testonly::abort_on_panic();
+    let ctx = &ctx::test_root(&ctx::AffineClock::new(5.));
+    Test {
+        network: Network::Real,
+        nodes: vec![(Behavior::Honest, 1), (Behavior::HonestNotProposing, 1)],
+        blocks_to_finalize: 10,
+    }
+    .run(ctx)
+    .await
+    .unwrap()
+}
